Fix sequence number assignment for EthSendTransaction
The sequence number setting used by web3 is a data race. To do this
requires the mechanism present in the mempool signing accounts package.

You should just send the transaction unsigned and Transactor will take
care of serialising sequence numbers. It looks like the only real issue
with this currently is mock key client only supports ed25519. I have
some vague memory of Sean improving this, should be easy to do.

Support ed25519 and secp256k1 in tests

Refactor away mock keys client replacing with MemoryKeyStore and
localKeyclient wrapper
