+++ integration/core/kernel_test.go
@@ -21,7 +21,6 @@ import (
 	"github.com/hyperledger/burrow/integration"
 	"github.com/hyperledger/burrow/integration/rpctest"
 	"github.com/hyperledger/burrow/keys"
-	"github.com/hyperledger/burrow/keys/mock"
 	"github.com/hyperledger/burrow/logging/logconfig"
 	"github.com/hyperledger/burrow/logging/loggers"
 	"github.com/hyperledger/burrow/rpc/rpctransact"
@@ -136,8 +135,8 @@ func bootWaitBlocksShutdown(t testing.TB, validator *acm.PrivateAccount, private
 		return err
 	}
 
-	kern.SetKeyClient(mock.NewKeyClient(privateAccounts...))
-	kern.SetKeyStore(keys.NewKeyStore(keys.DefaultKeysDir, false))
+	kern.SetKeyClient(keys.NewLocalKeyClient(keys.NewMemoryKeyStore(privateAccounts...), kern.Logger))
+	kern.SetKeyStore(keys.NewFilesystemKeyStore(keys.DefaultKeysDir, false))
 	ctx := context.Background()
 	if err = kern.Boot(); err != nil {
 		return err
